Refactor puffin_viewer to make it work in a browser
