+++ check.sh
@@ -7,7 +7,7 @@ set -eux
 cargo check --workspace --all-targets
 cargo test --workspace --doc
 cargo check --workspace --all-targets --all-features
-cargo check -p puffin_egui --lib --target wasm32-unknown-unknown --all-features
+cargo check -p puffin_viewer --lib --target wasm32-unknown-unknown --all-features
 cargo clippy --workspace --all-targets --all-features --  -D warnings -W clippy::all
 cargo test --workspace --all-targets --all-features
 cargo fmt --all -- --check
