+++ puffin_viewer/Cargo.toml
@@ -12,6 +12,9 @@ categories = ["development-tools::profiling", "gui"]
 keywords = ["profiler", "instrumentation", "gamedev"]
 include = [ "**/*.rs", "Cargo.toml"]
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+
 [dependencies]
 puffin_egui = { version = "0.11.0", path = "../puffin_egui" }
 puffin = { version = "0.12.0", path = "../puffin", features = ["packing", "serialization"] }
